Create arduino_temp_log.json
Modular Temperature Measurement and Logging

Measures temperature using thermistors and logs the data in csv format.

Designed to be modular in nature so that additional temperature sensors can be added easily, with the long term intent to expand the available sensor inputs or outputs.

The temperature information is extracted from a board of your choice (an Arduino board was originally used) using the Johnny-Five node.js library and the custom built Johnny-Five node-red node (available at https://www.npmjs.com/package/node-red-contrib-gpio, with details as to which micro controller boards are supported)

Can be used for any temperature monitoring application, but was built with the intent to log temperature data with a timestamp for research experiments. The data is logged in CSV format for easing importing into more suitable software packages for post-processing and analysis, such as MATLAB or Python.

A more detailed description of how to use this flow can be found at: https://sites.google.com/view/opensourcehardwaresouthafrica

update: 26Feb17: The flow has been modified to initialise the sensors in a seperate node, so that subsequent nodes can record information when promted by the flow. (Previously data recording nodes could not be setup to record data when prompted, as this created a new instance of the sensor each time, resulting in a oveflow of sensor instances).